Flock v0.1.0 Flock.Manager

This is the core module of Flock

It is in charged of keeping track of alive processes, communicating changes to other nodes, and receiving notifications of processes ending and updates from another node

Types

t()

t() :: %Flock.Manager{active: term(), local: list()}

Functions

alive?(name)

alive?(name :: Flock.worker_name()) ::
  boolean() |
  {:error, reason :: String.t()}

Returns if the process named name is alive using local node information

call(name, request)

call(name :: Flock.worker_name(), request :: term()) :: any()

Call to a worker

cast(name, request)

cast(name :: Flock.worker_name(), request :: term()) :: :ok

Cast to a worker

remove(name)

remove(name :: Flock.worker_name()) :: :ok | {:error, :not_found}

Removes a process on normal exit

start_link(opts)

Starts the Flock.Manager

start_worker(worker_spec)

start_worker(worker_spec :: Flock.worker_spec()) ::
  :ok |
  {:error, :already_exists}

Spawn a new process

stop(name)

stop(name :: Flock.worker_name()) :: :ok | {:error, :not_found}

Stops a process